requirements.txt
requests
dnspython
rich
cloudscraper
base64